% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cage}
\alias{cage}
\title{CAGE peak annotation from AnnotationHub}
\format{
A \linkS4class{GRanges} object.
}
\usage{
data(cage, package="crisprViz")
}
\description{
A sample of CAGE peak annotation from AnnotationHub (ascension number
AH5084), lifted over to hg38.
}
\keyword{datasets}
