% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cas9GuideSet}
\alias{cas9GuideSet}
\title{Example GuideSet targeting the human LTN1 gene.}
\format{
A \linkS4class{GuideSet} object.
}
\usage{
data(cas9GuideSet, package="crisprViz")
}
\description{
An example \linkS4class{GuideSet} object targeting the CDS of the human LTN1
gene. Generated using the SpCas9 CRISPR nuclease, the genome sequence
from the \code{BSgenome.Hsapiens.UCSC.hg38} package, and the gene model
from \code{txdb_human} in the \code{crisprDesignData} package (Ensembl
release 104).
}
\keyword{datasets}
