% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{datasets}
\alias{datasets}
\alias{krasGuideSet}
\alias{gpr21GuideSet}
\alias{mmp7GuideSet}
\alias{cage}
\alias{dnase}
\alias{cas9GuideSet}
\alias{cas12aGuideSet}
\title{Sample data sets}
\description{
Sample GuideSet and GRanges objects for demonstrative use in the vignette.
}
\keyword{datasets}
