% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dnase}
\alias{dnase}
\title{DNase I hypersensitive site annotation from AnnotationHub}
\format{
A \linkS4class{GRanges} object.
}
\usage{
data(dnase, package="crisprViz")
}
\description{
A sample of DNase I hypersensitive site annotation from AnnotationHub
(ascension number AH30743), lifted over to hg38.
}
\keyword{datasets}
