% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gpr21GeneModel}
\alias{gpr21GeneModel}
\title{CompressedGRangesList describing the human GPR21 gene.}
\format{
A \linkS4class{CompressedGRangesList} object.
}
\usage{
data(gpr21GeneModel, package="crisprViz")
}
\description{
A \linkS4class{CompressedGRangesList} object describing the gene model of
the human GPR21 gene. Coordinates were subset from \code{txdb_human} in
the \code{crisprDesignData} package (Ensembl release 104).
}
\keyword{datasets}
