% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gpr21GuideSet}
\alias{gpr21GuideSet}
\title{Example GuideSet targeting the human GPR21 gene.}
\format{
A \linkS4class{GuideSet} object.
}
\usage{
data(gpr21GuideSet, package="crisprViz")
}
\description{
An example \linkS4class{GuideSet} object targeting a section of the CDS of
the human GPR21 gene. Generated using the BE4max CRISPR nuclease, the
genome sequence from the \code{BSgenome.Hsapiens.UCSC.hg38} package, and
the gene model from \code{txdb_human} in the \code{crisprDesignData}
package (Ensembl release 104).
}
\keyword{datasets}
