% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ideogram_GRCm38}
\alias{Ideogram_GRCm38}
\title{Data.frame of mouse (ideogram) bands}
\format{
Data.frame with 5 columns:
\code{chrom}, \code{chromStart}, \code{chromEnd}, \code{name},
and \code{gieStain}.
}
\usage{
data(Ideogram_GRCm38, package="crisprViz")
}
\description{
A data.frame of mouse (ideogram) bands (hg38)
}
\details{
(how it was obtained, genome build, used in Gviz function to
plot IdeogramTrack for mouse
}
\keyword{datasets}
