% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{krasGeneModel}
\alias{krasGeneModel}
\title{CompressedGRangesList describing the human KRAS gene.}
\format{
A \linkS4class{CompressedGRangesList} object.
}
\usage{
data(krasGeneModel, package="crisprViz")
}
\description{
A \linkS4class{CompressedGRangesList} object describing the gene model of
the human KRAS gene. Coordinates were subset from \code{txdb_human} in
the \code{crisprDesignData} package (Ensembl release 104).
}
\keyword{datasets}
