% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mmp7GuideSet}
\alias{mmp7GuideSet}
\title{Example GuideSet targeting the human MMP7 gene.}
\format{
A \linkS4class{GuideSet} object.
}
\usage{
data(mmp7GuideSet, package="crisprViz")
}
\description{
An example \linkS4class{GuideSet} object targeting the promoter region of
the human MMP7 gene. Generated using the SpCas9 CRISPR nuclease, the
genome sequence from the \code{BSgenome.Hsapiens.UCSC.hg38} package, and
the gene model from \code{tss_human} in the \code{crisprDesignData}
package (Ensembl release 104).
}
\keyword{datasets}
