% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGuideSet.R
\name{plotGuideSet}
\alias{plotGuideSet}
\title{Plotting a GuideSet and other genomic annotation}
\usage{
plotGuideSet(
  guideSet,
  geneModel = NULL,
  targetGene = NULL,
  annotations = list(),
  from = NULL,
  to = NULL,
  extend.left = 0,
  extend.right = 0,
  bands = NULL,
  guideStacking = "squish",
  bsgenome = NULL,
  pamSiteOnly = FALSE,
  showGuideLabels = TRUE,
  onTargetScore = NULL,
  includeSNPTrack = TRUE,
  displayPars = NULL
)
}
\arguments{
\item{guideSet}{tbd}

\item{geneModel}{tbd}

\item{targetGene}{tbd}

\item{annotations}{tbd}

\item{from, to}{tbd}

\item{extend.left, extend.right}{tbd}

\item{bands}{tbd}

\item{guideStacking}{tbd}

\item{bsgenome}{tbd}

\item{pamSiteOnly}{tbd}

\item{showGuideLabels}{tbd}

\item{onTargetScore}{tbd}

\item{includeSNPTrack}{tbd}

\item{displayPars}{tbd (possible to implement?)}
}
\value{
A Gviz plot... see ?Gviz::plotTracks
}
\description{
Function to plot guide targets stored in a
\linkS4class{GuideSet} object in a gene browser view supported by
\code{\link[Gviz]}. Target gene isoforms and other genomic annotation,
along with the target chromosome ideogram and sequence, may also be
added, permitting a comprehensive visualization of the genomic context
targeted by spacers in the \linkS4Class{GuideSet}.
}
\examples{
\dontrun{
## WIP =====================================================================
}

}
\seealso{
\code{\link{plotMultipleGuideSets}}
}
\author{
Luke Hoberecht, Jean-Philippe Fortin
}
