% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{repeats}
\alias{repeats}
\title{Subset of repeat elements for hg38.}
\format{
A \linkS4class{GRanges} object.
}
\usage{
data(repeats, package="crisprViz")
}
\description{
A \linkS4class{GRanges} object describing a subset of repeat elements for
the hg38 genome. Coordinates were subset from \code{gr.repeats.hg38} in
the \code{crisprDesignData} package.
}
\keyword{datasets}
