% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMultipleGuideSets.R
\name{plotMultipleGuideSets}
\alias{plotMultipleGuideSets}
\title{Plotting a GuideSet and other genomic annotation}
\usage{
plotMultipleGuideSets(
  guideSets,
  geneModel = NULL,
  targetGene = NULL,
  annotations = list(),
  from = NULL,
  to = NULL,
  extend.left = 0,
  extend.right = 0,
  bands = NULL,
  bsgenome = NULL,
  onTargetScore = NULL,
  includeSNPTrack = TRUE,
  displayPars = NULL
)
}
\arguments{
\item{guideSets}{tbd}

\item{geneModel}{tbd}

\item{targetGene}{tbd}

\item{annotations}{tbd}

\item{from, to}{tbd}

\item{extend.left, extend.right}{tbd}

\item{bands}{tbd}

\item{bsgenome}{tbd}

\item{onTargetScore}{tbd}

\item{includeSNPTrack}{tbd}

\item{displayPars}{tbd (possible to implement?)}
}
\value{
A Gviz plot... see ?Gviz::plotTracks
}
\description{
Function to plot guide targets stored in a
\linkS4class{GuideSet} object in a gene browser view supported by
\code{\link[Gviz]}. Target gene isoforms and other genomic annotation,
along with the target chromosome ideogram and sequence, may also be
added, permitting a comprehensive visualization of the genomic context
targeted by spacers in the \linkS4Class{GuideSet}.
}
\examples{
\dontrun{
## WIP =====================================================================
library(crisprDesign)
data(guideSetExample)
library(crisprDesignData)
data(txdb_human)
gs <- guideSetExample[1:10]
targetGene <- "IQSEC3"
geneModel <- txdb_human
plotGuideSet(gs,
             geneModel=geneModel,
             targetGene=targetGene)

kras <- queryTxObject(txdb_human, "cds", "gene_symbol", "KRAS")
kras <- kras[1]
gs <- findSpacers(kras, crisprNuclease=SpCas9, bsgenome=BSgenome.Hsapiens.UCSC.hg38)
ideogram <- readRDS("inst/shiny/data/ideogram_hg38.rds")
}

}
\seealso{
\code{\link{plotMultipleGuideSets}}
}
\author{
Luke Hoberecht, Jean-Philippe Fortin
}
