% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMultipleGuideSets.R
\name{plotMultipleGuideSets}
\alias{plotMultipleGuideSets}
\title{Plotting a GuideSet and other genomic annotation}
\usage{
plotMultipleGuideSets(
  x,
  geneModel = NULL,
  targetGene = NULL,
  annotations = list(),
  from = NULL,
  to = NULL,
  extend.left = 0,
  extend.right = 0,
  margin = 1,
  includeIdeogram = TRUE,
  bands = NULL,
  bsgenome = NULL,
  onTargetScores = NULL,
  includeSNPTrack = TRUE,
  gcWindow = NULL
)
}
\arguments{
\item{x}{named list of GuideSets}

\item{geneModel}{tbd}

\item{targetGene}{gene symbol or id}

\item{annotations}{named list of annotations to plot (GRanges)}

\item{from, to}{see ?plotTracks}

\item{extend.left, extend.right}{see ?plotTracks}

\item{margin}{tbd}

\item{includeIdeogram}{boolean}

\item{bands}{see ?IdeogramTrack}

\item{bsgenome}{used in SequenceTrack}

\item{onTargetScores}{list of on-target scores}

\item{includeSNPTrack}{tbd}

\item{gcWindow}{tbd}
}
\value{
A Gviz plot... see ?Gviz::plotTracks
}
\description{
Function to plot guide targets stored in a
\linkS4class{GuideSet} object in a gene browser view supported by
\code{Gviz}. Target gene isoforms and other genomic annotation,
along with the target chromosome ideogram and sequence, may also be
added, permitting a comprehensive visualization of the genomic context
targeted by spacers in the \linkS4Class{GuideSet}.
}
\examples{
\dontrun{
## WIP =====================================================================
library(crisprDesign)
data(guideSetExample)
library(crisprDesignData)
data(txdb_human)
library(BSgenome.Hsapiens.UCSC.hg38)
gs <- guideSetExample[1:10]
targetGene <- "IQSEC3"
geneModel <- txdb_human
plotGuideSet(gs,
             geneModel=geneModel,
             targetGene=targetGene)

kras <- queryTxObject(txdb_human, "cds", "gene_symbol", "KRAS")
kras <- kras[1]
gs <- findSpacers(kras, crisprNuclease=SpCas9, bsgenome=BSgenome.Hsapiens.UCSC.hg38)
ideogram <- readRDS("inst/shiny/data/ideogram_hg38.rds")
}

}
\seealso{
\code{\link{plotMultipleGuideSets}}
}
\author{
Luke Hoberecht, Jean-Philippe Fortin
}
